+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Unreleased
+
+* Add `len` & `is_empty` functions
+
+# 0.17.0
+* (breaking) removed `Toast::font(font: FontId)`, this can now be done by using `egui::widget_text::RichText` and `RichText::font`. [#34]
+* Added support for `egui::widget_text::WidgetText` in Toasts, this allows for more customization of the text in the toast. [#34]
+
+[#34]: https://github.com/ItsEthra/egui-notify/pull/34
+
 # 0.16.0
 
 * (breaking) Updated to egui `0.29`.